services/web/requirements.txt
flask==2.0.2
Flask-SQLAlchemy==2.5.1
flask-migrate==3.1.0
psycopg2-binary==2.9.2

pytest==6.2.5
pytest-rabbitmq==2.2.0

pydantic==1.8.2
email-validator==1.1.3

pika==1.2.0
Flask-Mail==0.9.1